change socket

+++ BerkeleyServer/BerkeleyNetwork/include/PortHelper.h
@@ -8,8 +8,8 @@ namespace Network
         PortHelper() = delete;
         ~PortHelper() = delete;
 
-        inline static size_t TCP_PORT = 1001;
-        inline static size_t UDP_PORT = 1002;
+        inline static uint16_t TCP_PORT = 1001;
+        inline static uint16_t UDP_PORT = 1002;
     };
 }
 #endif //BERKELEYSERVER_PORTHELPER_H
